+++ sdk/resourcemanager/domainservices/armdomainservices/zz_generated_oucontaineroperations_client.go
@@ -50,7 +50,7 @@ func NewOuContainerOperationsClient(credential azcore.TokenCredential, options *
 
 // NewListPager - Lists all the available OuContainer operations.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-05-01
+// Generated from API version 2022-06-01
 // options - OuContainerOperationsClientListOptions contains the optional parameters for the OuContainerOperationsClient.List
 // method.
 func (client *OuContainerOperationsClient) NewListPager(options *OuContainerOperationsClientListOptions) *runtime.Pager[OuContainerOperationsClientListResponse] {
@@ -89,7 +89,7 @@ func (client *OuContainerOperationsClient) listCreateRequest(ctx context.Context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-05-01")
+	reqQP.Set("api-version", "2022-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil